Move `Sort` component to partial.
