+++ packages/template-retail-react-app/app/pages/product-list/partials/sort.jsx
@@ -1,0 +1,62 @@
+/*
+ * Copyright (c) 2023, Salesforce, Inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: BSD-3-Clause
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+import React from 'react'
+import PropTypes from 'prop-types'
+import {useHistory} from 'react-router-dom'
+import {useIntl} from 'react-intl'
+
+// Components
+import {Select, FormControl} from '@salesforce/retail-react-app/app/components/shared/ui'
+
+const Sort = ({sortUrls, productSearchResult, basePath, ...otherProps}) => {
+    const intl = useIntl()
+    const history = useHistory()
+
+    return (
+        <FormControl
+            aria-label={intl.formatMessage({
+                id: 'product_list.drawer.title.sort_by',
+                defaultMessage: 'Sort By'
+            })}
+            data-testid="sf-product-list-sort"
+            id="page_sort"
+            width="auto"
+            {...otherProps}
+        >
+            <Select
+                value={basePath.replace(/(offset)=(\d+)/i, '$1=0')}
+                onChange={({target}) => {
+                    history.push(target.value)
+                }}
+                height={11}
+                width="240px"
+            >
+                {sortUrls.map((href, index) => (
+                    <option key={href} value={href}>
+                        {intl.formatMessage(
+                            {
+                                id: 'product_list.select.sort_by',
+                                defaultMessage: 'Sort By: {sortOption}'
+                            },
+                            {
+                                sortOption: productSearchResult?.sortingOptions[index]?.label
+                            }
+                        )}
+                    </option>
+                ))}
+            </Select>
+        </FormControl>
+    )
+}
+
+Sort.propTypes = {
+    sortUrls: PropTypes.array,
+    productSearchResult: PropTypes.object,
+    basePath: PropTypes.string
+}
+
+export default Sort
